feat: restructure extension to support adhkar notifications and prayer time management
- Moved main entry point to src/extension.js
- Added new src/adhkar.js for handling adhkar HTML content and notifications
- Introduced src/cache.js for caching prayer times and managing state
- Created src/commands.js to register commands for configuring prayer times and opening adhkar panels
- Added src/constants.js for defining constants related to caching and prayer methods
- Implemented src/location.js for detecting user location and applying settings
- Developed src/notifications.js for managing status bar notifications and updates
- Included src/utils.js for utility functions related to date and configuration handling
